restore previous structure and cleanup
