+++ @server/@api-user/user.route.ts
@@ -1,16 +1,25 @@
 import express, { IRouter } from 'express';
 import {
+  getAllUsersController,
   getOneUserController,
+  deleteOneUserController,
   updateOneUserController,
-  deleteOneUserController
+  deleteAllUserController,
 } from '@server/@api-user/user.controller';
 import { UserRole } from '@server/@api-user/user.model';
 import { authenticateUserWithJWT, authorizeByUserRoles } from '@server/@api-auth/middlewares/auth.middleware';
 
 const router: IRouter = express.Router();
 
-router.get('/', authenticateUserWithJWT, authorizeByUserRoles([UserRole.User, UserRole.Admin]), getOneUserController);
-router.patch('/', authenticateUserWithJWT, authorizeByUserRoles([UserRole.User, UserRole.Admin]), updateOneUserController);
+router.get('/all', authenticateUserWithJWT, authorizeByUserRoles([UserRole.Admin]), getAllUsersController);
+router.get('/', authenticateUserWithJWT, authorizeByUserRoles([UserRole.Admin, UserRole.User]), getOneUserController);
+
+router.patch('/', authenticateUserWithJWT, authorizeByUserRoles([UserRole.Admin, UserRole.User]), updateOneUserController);
+
 router.delete('/', authenticateUserWithJWT, authorizeByUserRoles([UserRole.User]), deleteOneUserController);
+
+//-----------------------------------------------------//
+router.delete('/all', authenticateUserWithJWT, authorizeByUserRoles([UserRole.Admin]), deleteAllUserController);
+//-----------------------------------------------------//
 
 export { router };
