+++ tsconfig.json
@@ -37,7 +37,6 @@
     "paths": {
         "@trash/*": ["@server/@api-trash/*"],
         "@collabocate/*": ["@server/@api-manager_collabocate/*"],
-        "@admin/*": ["@server/@api-admin/*"],
         "@user/*": ["@server/@api-user/*"],
         "@auth/*": ["@server/@api-auth/*"],
         "@api-community/*": ["@server/@api-home/*"],
